fixed bugs from previous commits
