+++ miracle-messages/src/Actions/ForgotPasswordActions.js
@@ -1,27 +1,27 @@
-import axios from "axios"; 
+import axios from "axios"
 
-export const POST_VOLUNTEERS_SUCCESS
-export const POST_VOLUNTEERS_FAIL
+export const GET_VOLUNTEERS_SUCCESS = "POST_VOLUNTEERS_SUCCESS"
+export const GET_VOLUNTEERS_FAIL = "POST_VOLUNTEERS_FAIL"
 
 export const addEmail = email => async dispatch => {
-    dispatch({
-        type: GET_VOLUNTEERS_SUCCESS
-    })
+  dispatch({
+    type: GET_VOLUNTEERS_SUCCESS
+  })
 
-    axios
-        .post(
-            // API from Twilio goes here! 
-        )
-        .then(res => {
-            console.log("This is the response from AddEmail", res)
-            return dispatch({
-            type: GET_VOLUNTEERS_SUCCESS
-            })
-        })
-        .catch(error => {
-            console.log("This is the catch from AddEmail", error)
-            return dispatch({
-                type: GET_VOLUNTEERS_FAIL
-            })
-        })
+  axios
+    .post
+    // API from Twilio goes here!
+    ()
+    .then(res => {
+      console.log("This is the response from AddEmail", res)
+      return dispatch({
+        type: GET_VOLUNTEERS_SUCCESS
+      })
+    })
+    .catch(error => {
+      console.log("This is the catch from AddEmail", error)
+      return dispatch({
+        type: GET_VOLUNTEERS_FAIL
+      })
+    })
 }
