+++ miracle-messages/src/App.js
@@ -8,7 +8,8 @@ import LoginForm from "./Components/Login/LoginForm.js"
 import { PrivateRoute } from "./Components/PrivateRoute"
 import "./CSS/style.css"
 import LoginPage from "./Components/Login/UserLogin"
-import ConfirmPassword from "./Components/Login/ConfirmPassword"
+import ConfirmPassword from "./Components/Forms/ForgotPassword/ConfirmPassword"
+import ForgotPasswordForm from "./Components/Forms/ForgotPassword/ForgotPasswordForm"
 
 class App extends Component {
   render() {
@@ -26,7 +27,7 @@ class App extends Component {
           <Route
             exact
             path="/user/forgotpassword1"
-            component={ConfirmPassword}
+            component={ForgotPasswordForm}
           />
 
           <Route
